mrconvert: Avoid unnecessary warning
If -clear_property command_history is used in an mrconvert call, this will have an influence on the output even if there is no command_history entry present in the input image, as it will preclude the addition of the currently running command to the output header. As such, a warning regarding the absence of that field in the input image should not be issued.