DWI::Shells: Classify b=0 before clustering
Criterion of b=0 threshold should be applied to classify individual volumes as b=0 prior to the shell clustering algorithm being applied. This is necessary for the case where BValueEpsilon is greater than or equal to the b-value of the first non-b=0 shell, in order to prevent any b=0 volumes from being encapsulated within the smallest non-b=0 shell.