dw_scheme: Fix command-line import
In #2015, a change was made whereby the diffusion gradient table stored in the input image header would only be updated (in order to then propagate to output image headers) if the normalisation of diffusion gradient vectors had a substantial effect on the data. This however erroneously prevented imported gradient tables via the -grad or -fslgrad command-line options to be placed into the input header: DWI::get_DW_scheme() would return a valid matrix, but the input header would not be updated.