CohortDataImport::initialise(): Fix for cwd
Empty string output by Path::dirmane() if the location of the list file is the current working directory would lead to an erroneous Exception being thrown, as that string being empty is also used as a flag to indicate that none of the candidate directories were reliable references for finding all subject files.
This fix uses "." rather than "" for the current working directory, so that if a valid Exception is indeed thrown the error message should make more sense.