dwifslpreproc: Improve eddy error reporting
In cases where both cuda and openmp versions of eddy are available, the reason for the cuda version failing may be obscured because it is only the terminal output produced by the openmp version that is reported to the user. With this change, the messages produced by both cuda and openmp eddy commands will be concatenated into a new combined exception.