dwi2mask: Add config file options
The dwi2mask command is executed within some other MRtrix3 scripts. Now that it is implemented as a Python script with multiple algorithms to select from, it is necessary to be able to influence which dwi2mask algorithm is used whenever it is invoked from inside one of these scripts (and therefore the user does not have access via the command-line).