parse_ints(): Turn into a template
In order to support images with axis sizes greater than 2^31, a version of the parse_ints() function operating with 64-bit integers is required. This change turns this function into a template, and updates all usages accordingly. Note also that many usages of this function involve values that must be non-negative; for such instances the data handling is changed to unsigned integer type, and an Exception should be thrown during the string-to-unsigned-integer conversion, rather than relying on downstream data sanity checking.