run_tests: allow different location for MRtrix executables
This is to allow binary packages to be tested. To use, make sure you
have a clone of the MRtrix3 repo checked out, and the binary package you
want to test is installed. Then set the MRTRIX_BINDIR environment
variable to point to the location of the executables to be tested, and
invoke run_tests. For example:

MRTRIX_BINDIR=~/miniconda3/bin/ ./run_tests binaries