fix up looping constructs, etc
Note there's still an issue with the FFT shift here: this is due to the
indexing in the frequency domain, where zero frequency is at index 0,
and negative frequencies are in the range above N/2 (where N is the
length of the array). To get this right, you'd need to also fix up the
handling of this discontinuity at N/2.