Stats commands: Fix -strong option
Prior change in the impelementation of strong FWE control, changing the command-line option from receiving a boolean to being a flag, was not propagated to the point in code at which that option was read. This however now results in an error upon attempting to use that option due to merge of #1794, as it erroneously attempts to convert the next item on the command-line following the -strong option to a double.