maskfilter: Tweak to clean filter
Previously, those core voxels flagged for removal following erosion & connected component analysis would then be dilated in order to remove the full set of voxels deemed to be connected by a bridge smaller than or equal to the scale parameter.
This could however lead to the removal of voxels that are part of the main connected component and not of the bridged component flagged for removal, yet were within the range of the dilation operation.
This change constrains the removal of voxels from the mask based on per-iteration dilation, preventing that dilation from crossing voxels that are themselves not part of the original mask.