vectorstats: Permit numerical matrix data as input
For some experiments, it may be preferable for the primary input to the vectorstats command to be the raw data in matrix form, rather than a list of files (each of which contains numerical vector data) (which is consistent with the interface of other statistical inference commands). This change attempts data import based on the user-specified argument using both approaches.