dwi2mask trace: Experiment with iterative shell weighting
This is an experimental change to the trace algorithm. The original approach, where shell mean histograms are matched before combining them and performing automated thresholding, is still applied. However subsequently to this, each shell mean image is interrogated to determine how much it does or does not contribute to the separation of brain and non-brain, and a new set of weights are determined for the recombination of the shell means.