CI: Perform limited script testing
In "run_tests", rename previous "scripts" option to "scripts_full", and add new option "scripts_quick". The latter performs a manually curated limited set of tests, none of which have external neuroimaging software dependencies. This set of tests executes quickly enough to be included as part of Continuous Integration testing for merges to dev in addition to master; the tests are only executed on the MacOS workflow, as this consistently requires the least amount of time to complete.