dwifslpreproc: For EddyQC, attempt both Python versions
Originally, command "eddy_quad" was intentionally executed using the Python3 interpreter, as it was observed to not work with Python2. However more recently it has been reported from a user that explicitly using the Python3 interpreter results in failure, whereas invoking the script directly succeeds.
In order to cover off both cases, this commit modifies the dwifslpreproc script to attempt the "eddy_quad" command twice if necessary: once using a direct invocation, and once explicitly requesting the Python3 interpreter. A terminal warning will only be produced if the command is installed but both attempts fail.