Interp::Masked: Change implementation
In the original implementation, scanning through the dataset in order to produce an explicit binary voxel mask resulted in a noticeable delay in the commencement of tractography. With this change, the data in the voxel in which the point resides are instead checked for zero-filling at the point of the voxel() call.