dwi2mask trace: Move -volumes option to "dwi2mask mean"
For this algorithm, which simply performs a direct average of all DWI volumes (regardless of b-values, or number of volumes at each b-value), access as a command-line option to the "dwi2mask trace" algorithm was unusual, as very explicitly does not involve calculation of trace images. It has therefore been moved to its own dedicated algorithm.