dwi2mask: Exclude empty voxels from mask
When created, the dwi2mask trace algorithm included calculation of a mask where the input DWI data contains valid data, and multiplication of such with the final mask, to ensure that no voxels that do not include valid DWI data would be included in the final mask.
This functionality has now been moved to the dwi2mask wrapper script, so that this condition is applied regardless of the particular dwi2mask algorithm being used.