get_option_value(): Restrict usage
Function get_option_value() should only be used if accessing command-line options for which multiple uses are not permitted, and for which only a single Argument must be specified. If its use were to be permitted in other instances, it would potentially cause a misleading interpretation of the user's inputs.