Multithreading: Proposed alteration to functor duplication
When constructing multiple instances of a functor class in order to perform a processing task in a multi-threading fashion, query the functor class to see if it has a function called __clone(); if it does, use that function to generate the additional instances required, rather than any implicit or explicit copy-constructor. This will potentially enable applications where the functor to be multi-threaded is provided as a pointer to a base class, and thus appropriately generating clones of the derived class must be achieved through a virtual function.