dirorder: Fix segfault
Downstream of #1800, which introduced the functionality of determining the optimal reordering of directions for all possible selections of first volume, and choosing the optimal reordering from that set. Algorithm produces a segmentation fault if fewer than six input directions. This change forces the first volume in the input to be the first volume in the output in such scenarios (as there is no objective criterion by which to determine which direction is best to use first in this scenario).