dwifslpreproc: Compatibility between image padding and EddyQC
If using eddy slice-to-volume correction, and the DWI data are padded with an additional slice in the slice encoding direction for compatibility with the default topup config file, then EddyQC will fail due to the slspec file not containing a square matrix. To mitigate this, as well as make the results of EddyQC compatible with the output data of dwifslpreproc more generally, the effects of image padding applied within dwifslpreproc are undone prior to the execution of eddy_quad, including image outputs, text files that contain per-slice statistics, and the slspec file.
Closes #1781.