Attempted fixes to phase encoding handling
When importing data from / exporting data to phase encoding tables (whether in a basic tabular format or in the FSL eddy config / index format), potential axis permutation / flipping on image load / save needs to be taken into account, just as it is done when such data are stored in JSON files.

Addresses #2306.