DICOM: simplify handling of parsing in context of multi-frame
Changes to multiframe handling prompted by discussion on forum:
https://community.mrtrix.org/t/mrconvert-directions-in-0018-9089-not-detected/

The idea is to parse all tags regardless of which level in the hierarchy
they appear in, rather than only parsing when they're top-level or
within a Functional Group Sequence. Instead, we just need to ignore the
IconImageSequence, which contains tags that can conflict with the main
image; indeed, this seems to be the only reason the parsing worked as it
did so far.