dwifslpreproc: Handle non-ascii data in eddy error outputs
The FSL eddy commands sometimes contain a small block of binary data at the end of their outputs. If, when attempting to write those data to a text file in the scratchh directory, an encoding operation occurs, such data may yield a UnicodeEncodeError. This commit changes the output text files to be opened in binary mode, so such an encoding should no longer be applied.