tckgen: fix max angle exceeded at seed point with bidirectional tracking
With iFOD1, the second half of a bidirectional streamline starts in the
opposite direction from the first half. The seed direction is determined
first, and then used as the initial direction of tracking for both
halves. But the actual direction of the first step is taken as a random
sample within the angular constraint about that seed direction, and
because this happens both ways, and with the same seed direction, it is
possible for the angle between the two line segments either side of the
seed point to be up to twice the max angle.

The solution here involves setting the initial direction of tracking for
the second half of the streamline to the (opposite of the) *actual*
direction of the first step of the first half, rather than the seed
direction initially determined. This ensures that the first step of the
second half of the streamline is taken within the max angle constraint
of what is effectively the incoming direction of tracking.

To achieve this, the reverse_track() method was extended to perform the
actual track reversal and set the position and initial direction of
tracking, and overridden in the iFOD1 method to set things
appropriately.