DICOM: ugly hack to handle badly formatted Philips data
This is required to handle some Philips data, where the DW encoding
information is stored in 3 different locations: two with the official
DICOM tags (the latter with a private per-frame sequence), the other
using the Philips private tags. Unfortunately one of them is completely
incorrect (values stored as 7.23e+75 or some other equally nonsensical
value), but since it's the last one encountered on file, it would be the
value recorded without this hack.