dwifslpreproc: Fix applytopup call
Due to potential permutation of phase encoding direction on export of eddy-style config / index files (#2308), selection of volumes based on the internally-stored phase encoding table via dwiextract could fail.
This change instead performs the secltion of voplumes to be fed to applytopup based on the index file.
Echoes change proposed in 76f52505 as part of #2348, but is a requisite fix for master (as part of #2287).