fixelconnectivity: Use normal image handling for write
Remove custom code for exporting fixel-fixel connectivity matrix to filesystem, which was utilised because the total number of connections was not known a priori. Instead, explicitly calculate the number of supra-threshold connections, and then use the back-end library image access functionalities for writing these data. Additionally try to catch the case where the output destination is unable to support the volume of data to be written, and produce a more sensible error message.