mrtrix3.run: Fix default MRTRIX_LOGLEVEL
Default value for environment variable MRTRIX_LOGLEVEL was being erroneously written as an integer type rather than a string. In some environments, the subprocess module attempts to check whether environment variabes correspond to filesystem paths, and this results in a TypeError due to this specific envvar not being a string.