tcksift2: Fix masking of fixels with one streamline
The mechanism implemented for omitting exclusion of fixels traversed by only one streamline was erroneously being incremented based on the number of threads used to calculate the track density in each fixel. This resulted in this mechanism having no effect when executed with multi-threading, and an inability for the command to do anything if executed without multi-threading as all fixels would be excluded.