mrview: add support for custom colourmaps
This adds support for loading custom colourmaps from file at runtime.
This addition will scan through the folder ~/.mrtrix/colourmaps/ (if it
exists) and load each file as a distinct colourmap, making it available
on the GUI using the name of the file as the colourmap name. The format
of each colourmap file is as a matrix of 4 columns, with each column
corresponding to:

[ control_point red green blue ]

The first and last row control point values should be zero & one
respectively, and all intermediate control points should increase
monotonically between these values. For example:

0 0 0 0
0.5 0 0 1
1 1 0 0

increases from black to pure blue at 50%, then to pure red at 100%. For
intermediate intensities, the colour is derived by linear interpolation
from the nearest control points.