DICOM: fix import of gradient table
This should avoid problems when the DICOM files contain both GE-specific
entries (which need to be interpreted relative to the slice axes) and
DICOM standard entries (which need to be interpreted relative to the
scanner axes). This defaults to using the DICOM standard entries if
present, falling back to the GE-specific ones if found.