dwi2tensor: add small value to ADC squared when computing kurtosis components
This is to avoid instabilities in the estimated kurtosis values when the
estimated ADC is essentially zero, which can easily happen when
non-negativity constraints are applied.

In testing, this makes no detectable change to the estimated kurtosis
values (for those voxels with a sensible estimate), and no difference at
all to all other parameters.