GitHub Actions: fix clang warnings
See also previous conversation on the topic in #2392, in particular
commit aa5a354. The approach here is simply to ditch all the unused
variables and associated code identified as having no effect. But it may
be desirable to actually fix the code to use the variables as might have
originally been intended, particularly for tckedit. I'll push updates to
that effect in a subsequent commit.