tckgen: New algorithm "BallSticks"
This algorithm is intended to behave in a similar way to FSL's probtrackx, though is certainly not guaranteed to be a perfect match. This may be useful for experiments comparing specifically tractography algorithms, where other variables need to remain fixed.
A notable back-end change is that file src/dwi/bootstrap.h has been moved to src/dwi/tractography. This makes sense given that the functionality it provides is specifically around drawing unique bootstsrap realisations along a streamline trajectory where the set of voxels from which data will be required is not known a priori. In addition, there are now two different mechanisms for drawing such realisations within the tractography framework. The existing mechanism used for the tensor_prob algorithm, where a functor (WildBootstrap in that case) is used to generate a new data realisation based on the empirical data, remains but is renamed as BootstrapGenerate. A new mechanism, BootstrapSample, is used where multiple realisations of a model have been generated a priori, and are provided as an additional dimension within the image data.