dwibiasnormmask: New config file option "DwibiasnormmaskMaskAlgorithm"
This config file option will allow users to specify the default algorithm that will be used wthin the iterative process to derive the brain mask. This is a different step to derivation of an initial brain mask for response function estimation / MSMT CSD / mtnormalise, for which dwi2mask must be used. Rather, during the iterative process, one has access to the ODF tissue sum image, which may be used for brain masking, and so it may be preferable for something other than dwi2mask to be used taking that image as input.