DWI::Directions::FastLookupSet: Change implementation
Previous implementation was based on a naive exhaustive search of a grid of points within each search grid element. This failed to make use of the direction adjacency information available. New implementation does a gradient descent on directions based on this adjacency. The "fast" version is based on using the spherical coordinates of the input direction to access pre-calculated nearest directions in order to better seed this gradient descent.