Remove ghosting of MR::Registration::fod_options()
There were two versions of this function, but the ghosting was not discovered because cmd/mrregister.cpp was not including src/registration/shared.h. It was therefore using a duplicate that was defined in src/registration/linear.h, which is not really appropriate given that FOD reorientation is most commonly associated with non-linear registration.