fod2fixel: Write lookup image during processing
Rather than storing the LUT in RAM for every processed voxel and only writing the lookup image at command completion, fill the output lookup image with these contents as each voxel is processed. Unlike the index image and fixel data files, for the lookup image it is not necessary to know the total number of fixels at the point of image creation.