fod2fixel: Permit attributing one dixel to multiple fixels
Where a dixel resides at a trough between two lobes, previously that same would eventually be assigned to the fixel with the largest peak amplitude. With this change, those samples will instead be split evenly between the adjacent fixels, contributing a reduced amount to their respective integrals and direction calculations, and will be present in the direction masks for all such fixels.