Fixel format: Dixel masks are mutually exclusive
Partial reversion of e856b3e2.
While allowing a single direction to be ascribed to multiple fixels is a neat idea, and theoretically possible, it introduces all sorts of subsequent complexities, which are probably not worth the difficulty.
By making these masks mutually exclusive, not only can the retrospective assignment mechanism within the FMLS segmenter be simplified, but things like assignment of streamlines to fixels are much cleaner.