dir* command modifications for fixed direction handling
Modifications to dirgen, dirflip and dirorder.
These modifications were made for the purpose of embedding a set of fixed directions at the head of a gradient table scheme, and only optimising what is to come after that set. This is a minimal set of changes required to achieve that purpose; while there are other commands that could hypothetically undergo comparable changes, these were only done for a specific experiment, and features such as multi-shell or multi-phase-encode handling were deemed to be unnecessary.