tckmap: Fix to fixel mapping
Need ability to distinguish between the header that defines the output image, and the header that defines the voxel grid to which the streamlines are mapped, since for fixels these are drastically different. While existing variable "template_header" theoretically could have been re-used for this purpose, I chose to leave that fixed as what was read from the input to the -template option if used and invalid if not used.