DWI::Fixel_map: Fix FMLS LUT
Due to changes in #2639, the DWI::FMLS class no longer populates the fixel lookup table per voxel. However the DWI::Fixel_map class was still dependent on this information. In this commit, the DWI::Fixel_map class constructs that lookup table locally in order to facilitate compilation processing. Note however that results will not be identical to that of version 3.0.x, since this version of the LUT construction does not include dilation to allocate all directions to the nearest fixel, as is the default behaviour in that version of the software. Further, this change will become redundant with the changes in #2657.