First commit of overhauled fixelcorrespondence and new fixel2fixel command
Previous fixelcorrespondence command served a dual purpose: it would both determine the appropriate subject fixel to use in projecting data to a given template fixel, and then perform that projection to produce subject data for the template fixel dataset. This changeset breaks this process into two separate steps. The fixelcorrespondence command now produces as its output a dataset that represents the correspondence mapping between subject and template fixels. New command fixel2fixel is then responsible for utilising that pre-determined mapping to project data from subject fixels to template fixels.
The separation of these stages permits the exploration of alternative strategies for establishing the correspondence between subject and template fixles, inclusing accounting for cases where that correspondence is not a trivial one-to-one mapping.