dwi2mask consensus: Generalise -algorithm interface
Allow user input to -algorithm command-line option to be either a space-separated list (which involves argparse consuming multiple arguments), or a single string as a comma-separated list (which is more consistent with the rest of MRtrix3).
In addition, do not prevent execution due to the absence of template image information if there is no algorithm included in the list to execute that requires the use of such image data.