dwicat: Support different voxel grids
The dwicat script now serves two key purposes. In addition to estimating a global intensity shift between two or more series based on the aggregate b=0 intensity, it will now also interrogate the voxel grids of the input images, and decide how best to combine those data; whether it be straight concatenation, rigid-body transformation to a central unbiased header, or resampling all data onto a new common voxel grid if necessary.
Other changes to facilitate this / made in the process:
- dwicat new command-line option -nointensity to disable the intensity matching functionality of the script.
- mraverageheader: -resolution command-line option removed, and replaced with -spacing. There are now four options to choose from rather than two, and "resolution" was no longer the appropriate term to disambiguate them / describe the mechanism. In particular, the ability to set the voxel spacings based on either the minimum or mean of the input image axes closest to the average header axes, means that in the scenario where one has two DWI series with the same voxel spacing but different affine transforms, one can calculate an average voxel grid that lies between the two input voxel grids but has precisely the same voxel spacing, such that it is not necessary to perform interpolation to get the data from both series onto that common grid.
- transformcalc: Tweaked example description of "header" function.
Axes::get_permutation_to_make_axial(): Renamed to get_shuffle_to_make_axial() as the result is not constrained to axis permutations only, and also broke calculation of specifically the permutations into its own function as that could be used for the new functionalities in mraverageheader.