dwicat: SUpport different voxel grids
This commit is a manual cherry-pick of commit 1203d745dcc19be811e28b0a4f53fa16566204f0, which is on the dev branch as at 3.0.4 (ie. intended for 3.1.0), but applied to the status of the master branch at time of creation, which is post 3.0.4.
This is to provide this functionality to external tools without bringing in additional MRtrix3 changes that exist on the development branch.