Handle QFileOpenEvent using custom event handler in mrview
Previously, mrview required special handling of QFileOpenEvent and thus bool QApplication::event(QEvent*) was defined mrview.cpp. This strategy is no longer possible since now the code in src is built as a shared library, which requires the function to be defined in the shared library. To get around this we instead rely on a custom event handler using std::function.