Python: Initial work toward wrappers for executables
Change filesystem structure of python/.
No need to preserve the lib/mrtrix3/ sub-directory structure in the repository; that can be constructed exclusively in the build directory.
Executables now reside in sub-directory python/scripts/.
For algorithm-based scripts, all code is placed in a sub-directory of python/scripts/, with the previous bin/ file contents now placed in __init__.py of that directory, thus co-locating all relevant source code for such commands.
The algorithm module is simplified somewhat since the algorithm files are co-located with the interface source file.
Delete file python/bin/mrtrix3.py, and script source files no longer load the mrtrix3 module and invoke the execute function at the end of the file; an alternative mechanism for loading the API and script entrypoint will be created subsequently to this commit.