dwifslpreproc: Fix for PE direction, but not readout time, in header
If the phase encoding information present in the header is only the phase encoding direction, but not readout time is present, then the constructed phase encoding matrix will possess only three columns, not four. If this happens, but an attempt is then made to cross-reference the readout time in the phase encoding table generated from header content against what the user has specified at the command-line, then an unhandled exception will occur. This change prevents specifically the comparison of readout times from being made if no such information is present.
Replacement for erroneous commit faf0788a21c80498d431ed017661c0c179de2437.
Also includes additional test that results in an unhandled exception on 3.0.4 but executes successfully with this change.