run.command(): Do not hide list inputs
Where a command receives a long list of inputs, while collapsing that set may be preferable from the perspective of user terminal information, it should not be hidden from a raw log of the set of commands executed.
This change is not currently of high consequence as there is no mechanism that reads from that log file.