generate_interfaces.sh: Remove legacy .txt file extension
Files ending with .txt are ignored by git due to the contents of .gitignore.
To get around this, interface text files are written with no file extension.
This however would have resulted in a collision between the interface files generated for Python script algorithm interfaces, and the sub-directories for the algorithms contained within.
To get around thhis, files for the former are written into their own sub-directories.