configure: Enforce C++17
Construction of a 3.1.0 release involves taking features on the development branch and separating them from the adoption of cmake in #2689. Some of these changes involve the use of C++17 features. However the transition to defaulting to C++17 was performed exclusively within the cmake ecosystem. This commit modifies the legacy "configure" script to reflect the increment to use of C++17.