Python API: Change citation notification
Issuing a general citation warning is something that MRtrix3 Python commands do, but MRtrix3 C++ binaries do not. This may undesirably dilute the importance of recognition of utilisation of non-MRtrix3 commands, which previously modulated the content of this message. With this change, the citation warning is only issued if there is at least one external citation.