Python CLI: Default flags to None
In the C++ CLI, there are command-line options that themselves receive no arguments. Code typically tests for the mere presence of that option having been specified at the command-line interface. The python argparse module does not offer a direct translation of such. This change modifies the relevant behaviour to be somewhat more faithful, in that if the user does not specify that command-line option, the value stored will be None.
(Without this change, typically the value of False is stored. However with the introduction of command-line options that take as input a boolean value, which could be True or False as specified by the user or None if not specified at all, it would be more consistent with such if command-line flags store either True or None, rather than True or False. It also means that app.ARGS can be filtered according to which entries contain None to know what command-line options the user actually specified.