dwishellmath: Enhancements
Previously, alterations to the gradient table were evaluated exclusively based on axis permutations and flips, in either the real / scanner space or bvec basis. With this change, the command can additionally consider the prospect where the gradient table has been erroneously interpreted as containing coefficients in one basis when in fact those values should have been interpreted in the other basis. This error can be in addition to possible axis permutations and/or flips in one of the two bases.
Further, the command will now yield a non-zero return code if it is determined to be likely that the input gradient table is inappropriate.