Fix compilation errors with C++20
PerThread and Shared in threaded_loop.h now longer use references as
data members. This fixes some errors when compiling with C++20 as the
minimum standard (due to errors in initialising the variables in the
constructors) and also allows us to avoid following the rule of 5 (e.g.
no need to define custom copy/move constructors and assignment
operators).