Add interface library for project-wide compile options
Previously we were relying on the use of add_compile_options and
add_compile_definitions for defining compiler definitions and options for
the entire project. This commit replaces that behaviour by defining a new
mrtrix::common INTERFACE library that links to all targets in the
project (the library is linked to mrtrix::core). This is more idiomatic
and prevents the propagation of compiler definitions to targets that do
not link to mrtrix::core (useful when compiling MRtrix3 as a subproject
of anothe project).