run.command(): Change to shebang handling
If the command being run is an MRtrix3 executable (as established by the construction of that list at build time), and it is written in Python, then use the same Python interpreter as that currently executing. In any other circumstance, honour the shebang as specified in the command being invoked.