Disable the use of LLD by default
We no longer use LLD by default if it is available for reasons mentioned in #2986.
The use of LLD can now be enabled by using -DMRTRIX_USE_LLD=ON.
Additionally, we use CMake's check_cxx_compiler_flag to see if the compiler supports
LLD regardless of the platform (instead of simply relying on the presence of ld.lld).