tcksift2: Fix streamline exclusion for low weights
The documentation indicates that if a streamline obtains a weighting coefficient that is equal to or lesser than the minimum permissible, then it should be removed from the optimisation altogether; that is, given a weighting coefficient of -inf and a weighting factor of 0. However looking at the code it seems that this was not being applied.