Improve dash checking for option argument warnings
Replaced the use of `is_dash()` with a new `starts_with_dash()`
for the initial check in the `ParsedOption` constructor.
The previous `is_dash()` only returned true if the entire argument
string was a single dash character. This meant that option arguments
like "-foo" (which could be mistaken for another option) were not
triggering the intended warning if the option type did not accept a
leading dash (e.g., for negative numbers).