Trial fixes to external phase encoding table handling
Testing suggests that when the FSL software interprets phase encoding information stored in text files (whether in the "topup" format, with one row per volume, or in the "eddy" format, with one row per unique phase encoding and a second file indexing each volume into rows of that table), the interpretation of phase encoding directions is with respect to the FSL internal image convention rather than as the image is stored on disk, similarly to as occurs for diffusion gradient tables in the bvecs/bvals format. This commit is a first attempt at implementing proper support for these formats, faciliating their external testing. Of note is the behaviour of the -import_pe_* and -export_pe_* options. The -*_pe_table options retain their prior operation (with the exception that they now write the generating command as a comment at the header of the file). The -*_pe_eddy options will hopefully now appropriately deal with the flipping of the first axis where relevant. New options -*_pe_topup are very similar to the pre-existing -*_pe_table options, except that the first axis is flipped when required for proper interpretation by the FSL software.